[BE] refactor #93: 패키지명 단수로 통일 및 공백 컨벤션 통일
