+++ backend/d102-common/src/main/java/com/d102/common/exception/ExceptionType.java
@@ -14,7 +14,7 @@ public enum ExceptionType {
     private final String code;
     private final String message;
 
-    ExceptionType (String code, String message) {
+    ExceptionType(String code, String message) {
         this.code = code;
         this.message = message;
     }
