+++ backend/d102-common/src/main/java/com/d102/common/exception/custom/ConflictException.java
@@ -6,7 +6,7 @@ import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
-public class ConflictException extends RuntimeException{
+public class ConflictException extends RuntimeException {
 
     private final ExceptionType exceptionType;
