+++ backend/d102-common/src/main/java/com/d102/common/util/RegExpressions.java
@@ -1,4 +1,4 @@
-package com.d102.common.utils;
+package com.d102.common.util;
 
 public class RegExpressions {
     public static final String strongPassword = "^(?=.*[A-Za-z])(?=.*\\d)(?=.*[!@#$%^&*()_+\\-=?\\{\\};':\"|,.<>/~`\\[\\]\\\\])[A-Za-z\\d!@#$%^&*()_+\\-=?\\{\\};':\"|,.<>/~`\\[\\]]{8,}$";
